requirements.txt
argparse
wget
pdfminer.six
python-Levenshtein
